+++ .babelrc
@@ -1,10 +1,10 @@
 {
   "presets": [
-    "es2015",
-    "react",
-    "stage-3"
+    "@babel/env",
+    "@babel/react",
+    "@babel/stage-3"
   ],
   "plugins": [
-    "transform-runtime"
+    "@babel/transform-runtime"
   ]
 }
