+++ .vscode/settings.json
@@ -1,0 +1,15 @@
+{
+  "editor.rulers": [80],
+  "editor.tabSize": 2,
+  "search.exclude": {
+    "**/node_modules": true,
+    "**/dist": true
+  },
+  "files.exclude": {
+    "**/.git": true
+  },
+  "[json]": {
+    "editor.formatOnSave": false
+  },
+  "editor.formatOnSave": true
+}
