+++ package.json
@@ -6,17 +6,26 @@
   "repository": "git@github.com:artsy/stitch.git",
   "license": "MIT",
   "scripts": {
-    "build": "yarn clean:build && babel lib --ignore lib/__tests__ --out-dir dist",
+    "build": "yarn clean:build && babel src --ignore src/__tests__ --out-dir dist",
     "clean:build": "rm -rf dist",
     "coverage": "jest --coverage",
     "prepublishOnly": "yarn build",
     "test": "jest",
     "test:watch": "jest --watch"
   },
+  "prettier": {
+    "arrowParens": "always",
+    "bracketSpacing": true,
+    "semi": false,
+    "singleQuote": true
+  },
   "jest": {
     "setupFiles": [
       "./.test-setup"
-    ]
+    ],
+    "transform": {
+      ".(js|jsx)": "babel-jest"
+    }
   },
   "authors": [
     "Christopher Pappas <christopher@artsymail.com>",
@@ -24,23 +33,26 @@
     "Artsy <https://artsy.net>"
   ],
   "dependencies": {
-    "babel-runtime": "^6.26.0",
+    "@babel/runtime": "^7.0.0-beta.40",
     "consolidate": "^0.14.5",
     "lodash": "^4.17.4"
   },
   "peerDependencies": {
-    "react": "^15.0.0-0 || ^16.2.0-0",
-    "react-dom": "^15.0.0-0 || ^16.2.0-0",
-    "styled-components": "^2.2.3"
+    "react": "^15 || ^16",
+    "react-dom": "^15 || ^16",
+    "styled-components": "^2 || ^3"
   },
   "devDependencies": {
     "@artsy/express-reloadable": "^1.0.6",
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^7.2.3",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-3": "^6.24.1",
+    "@babel/cli": "^7.0.0-beta.40",
+    "@babel/core": "^7.0.0-beta.40",
+    "@babel/plugin-transform-runtime": "^7.0.0-beta.40",
+    "@babel/preset-env": "^7.0.0-beta.40",
+    "@babel/preset-react": "^7.0.0-beta.40",
+    "@babel/preset-stage-3": "^7.0.0-beta.40",
+    "babel-core": "^7.0.0-0",
+    "babel-eslint": "^8.2.2",
+    "babel-jest": "^22.4.1",
     "ejs": "^2.5.7",
     "eslint": "^4.5.0",
     "eslint-config-standard": "^10.2.1",
@@ -52,7 +64,8 @@
     "eslint-plugin-standard": "^3.0.1",
     "handlebars": "^4.0.10",
     "jade": "^1.11.0",
-    "jest": "^20.0.4",
+    "jest": "^22.4.2",
+    "prettier": "^1.10.2",
     "pug": "^2.0.0-rc.3",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
